OT98-62: move template into template package and make it inmutable
