+++ src/main/java/com/alkemy/ong/common/mail/template/ContactTemplateEmail.java
@@ -1,21 +1,19 @@
-package com.alkemy.ong.model.request;
+package com.alkemy.ong.common.mail.template;
 
 import com.alkemy.ong.common.mail.IContent;
 import com.alkemy.ong.common.mail.IEmail;
-import lombok.Getter;
-import lombok.Setter;
 
-@Getter
-@Setter
 public class ContactTemplateEmail implements IEmail, IContent {
 
   private static final String TYPE = "text/plain";
-
   private static final String VALUE = "Thank you for contact us";
-
   private static final String SUBJECT = "Contact Receive";
 
-  private String email;
+  private final String email;
+
+  public ContactTemplateEmail(String email) {
+    this.email = email;
+  }
 
   @Override
   public String getType() {
